requirements.txt

#numpy==1.21.1
#pandas==1.3.1
#requests==2.26.0
#streamlit==1.0.0
#waitress==2.1.0
#fastapi==0.115.6
#google-generativeai==0.8.3
#transformers==4.47.0
#torch==2.5.1


numpy
pandas
requests
streamlit
joblib
gunicorn
fastapi
google-generativeai
fastapi[standard]
transformers 
torch 